refactor : RecordService 클래스 신설하여 역할 분리
